+++ lib_/_conffiles_.py
@@ -20,6 +20,8 @@ import src.conffiles
 # built-in python imports
 # no imports
 
+__LIB_VERSION__ = 1
+
 # constants
 CONFIG_DIRECTORY = String(src.conffiles.CONFIG_DIRECTORY, *default_pos())